* Contrib: Added an updated version of the configuration file merger tool.
This one lets you choose which value to keep between the two configuration files.
Make sure to carefully read the file's lines 3 to 9 to make sure you don't misuse it and get back at me later raging.
It is not exactly as I would like it to be, but current state will have to do.

You can try it out live at: http://tinyurl.com/kzxmrlm